fix: incorrect equal
