+++ packages/mask/src/extension/popups/pages/Personas/Accounts/index.tsx
@@ -1,47 +1,21 @@
-import { memo, useCallback, useMemo } from 'react'
+import { memo, useCallback } from 'react'
 import { useTitle } from '../../../hook/useTitle'
 import { AccountsUI } from './UI'
 import { PersonaContext } from '../hooks/usePersonaContext'
 import { useI18N } from '../../../../../utils'
-import { compact, isEqual, unionWith } from 'lodash-unified'
-import { EMPTY_LIST, EnhanceableSite, PopupRoutes, ProfileIdentifier } from '@masknet/shared-base'
+import { compact } from 'lodash-unified'
+import { EnhanceableSite, PopupRoutes } from '@masknet/shared-base'
 import { useAsyncFn } from 'react-use'
 import Services from '../../../../service'
 import type { Account } from '../type'
 import { useNavigate } from 'react-router-dom'
-import { NEXT_ID_PLATFORM_SOCIAL_MEDIA_MAP } from '@masknet/shared'
 import { getEnumAsArray } from '@dimensiondev/kit'
 
 const Accounts = memo(() => {
     const { t } = useI18N()
 
     const navigate = useNavigate()
-    const { proofs, currentPersona, setSelectedAccount } = PersonaContext.useContainer()
-
-    const accounts = useMemo(() => {
-        if (!currentPersona) return EMPTY_LIST
-
-        const localProfiles = currentPersona.linkedProfiles.map<Account>((profile) => ({
-            ...profile,
-            identity: profile.identifier.userId,
-        }))
-
-        if (!proofs) return localProfiles
-
-        const remoteProfiles = proofs
-            .filter((x) => !!NEXT_ID_PLATFORM_SOCIAL_MEDIA_MAP[x.platform])
-            .map<Account>((x) => {
-                return {
-                    ...x,
-                    identifier: ProfileIdentifier.of(
-                        NEXT_ID_PLATFORM_SOCIAL_MEDIA_MAP[x.platform],
-                        x.identity,
-                    ).unwrap(),
-                }
-            })
-
-        return unionWith(localProfiles, remoteProfiles, (a, b) => isEqual(a.identity, b.identity))
-    }, [proofs, currentPersona])
+    const { currentPersona, setSelectedAccount, accounts } = PersonaContext.useContainer()
 
     const definedSocialNetworks = compact(
         getEnumAsArray(EnhanceableSite).map((x) => {
