+++ packages/mask/src/extension/popups/pages/Personas/hooks/usePersonaContext.ts
@@ -1,15 +1,16 @@
 import { createContainer } from 'unstated-next'
 import { useValueRef } from '@masknet/shared-base-ui'
-import { ECKeyIdentifier, EMPTY_LIST, PersonaInformation } from '@masknet/shared-base'
+import { ECKeyIdentifier, EMPTY_LIST, PersonaInformation, ProfileIdentifier } from '@masknet/shared-base'
 import { currentPersonaIdentifier } from '../../../../../settings/settings'
 import { useAsyncRetry } from 'react-use'
 import Services from '../../../../service'
-import { head } from 'lodash-unified'
-import { useCallback, useEffect, useState } from 'react'
+import { head, isEqual, unionWith } from 'lodash-unified'
+import { useCallback, useEffect, useMemo, useState } from 'react'
 import { MaskMessages } from '../../../../../utils'
 import { NextIDProof } from '@masknet/web3-providers'
 import type { Account } from '../type'
 import { initialPersonaInformation } from './PersonaContextInitialData'
+import { NEXT_ID_PLATFORM_SOCIAL_MEDIA_MAP } from '@masknet/shared'
 
 function useSSRPersonaInformation() {
     const [personas, setPersonas] = useState(initialPersonaInformation)
@@ -50,7 +51,35 @@ function usePersonaContext() {
         }
     }, [currentPersona])
 
+    const accounts = useMemo(() => {
+        if (!currentPersona) return EMPTY_LIST
+
+        const localProfiles = currentPersona.linkedProfiles.map<Account>((profile) => ({
+            ...profile,
+            identity: profile.identifier.userId,
+        }))
+
+        if (!proofs) return localProfiles
+
+        const remoteProfiles = proofs
+            .filter((x) => !!NEXT_ID_PLATFORM_SOCIAL_MEDIA_MAP[x.platform])
+            .map<Account>((x) => {
+                return {
+                    ...x,
+                    identifier: ProfileIdentifier.of(
+                        NEXT_ID_PLATFORM_SOCIAL_MEDIA_MAP[x.platform],
+                        x.identity,
+                    ).unwrap(),
+                }
+            })
+
+        return unionWith(remoteProfiles, localProfiles, (a, b) =>
+            isEqual(a.identity?.toLowerCase(), b.identity?.toLowerCase()),
+        )
+    }, [proofs, currentPersona])
+
     return {
+        accounts,
         selectedAccount,
         setSelectedAccount,
         avatar,
